Refactored the search path and function Refactored so that search is in it's own file Relates #42
