+++ src/controllers/getresources.js
@@ -1,15 +1,15 @@
 const getData = require('../queries/getData.js');
 
-module.exports = (req, response) => {
+module.exports = (req, res) => {
   getData()
-    .then(res => {
-        let output = JSON.stringify(res.rows);
-        response.writeHead(200, {
-          'content-type': 'application/json'
-        });
-        response.end(output);
-      })
-      .catch(err => {
-        console.log('error with getData in router');
-      })
+    .then(results => {
+      let output = JSON.stringify(results.rows);
+      res.writeHead(200, {
+        'content-type': 'application/json'
+      });
+      res.end(output);
+    })
+    .catch(err => {
+      console.log('Error with getData in router');
+    })
 }
