+++ src/controllers/search.js
@@ -1,0 +1,17 @@
+const searchData = require('../queries/searchData.js');
+
+module.exports = (req, res) => {
+  const searchquery = req.url.split('?')[1];
+  searchData(searchquery)
+    .then(results => {
+      let output = JSON.stringify(results.rows);
+      console.log("I am response from searchData: ", results.rows);
+      res.writeHead(200, {
+        'content-type': 'application/json'
+      });
+      res.end(output);
+    })
+    .catch(err => {
+      console.log("error with searchData in router");
+    })
+}
